@@ -4,13 +4,12 @@ version: 0.0.2
 homepage: https://github.com/smartbackme/flutter_interceptor
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
-  flutter: ">=1.17.0"
+  sdk: '>=3.0.6 <4.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
-  dio: ^4.0.0
+  dio: "^5.2.0+1"
 
 
 dev_dependencies:
