@@ -1,14 +1,10 @@
-import 'dart:convert';
+import 'dart:convert' as convert;
 
 import 'package:dio/dio.dart';
 import 'package:flutter_interceptor/src/model/http_transaction.dart';
-import 'dart:convert' as convert;
-
 import 'package:flutter_interceptor/src/tools/interceptor_manager.dart';
 
 class UiNetInterceptor extends Interceptor {
-
-
   @override
   void onRequest(RequestOptions options, RequestInterceptorHandler handler) {
     InterceptorManager.instance.onRequestTime(options.hashCode, DateTime.now());
@@ -17,29 +13,34 @@ class UiNetInterceptor extends Interceptor {
 
   @override
   void onResponse(Response response, ResponseInterceptorHandler handler) async {
-    InterceptorManager.instance.onResponseTime(response.requestOptions.hashCode, DateTime.now());
-    InterceptorManager.instance.onSave(setData(response, response.requestOptions));
+    InterceptorManager.instance
+        .onResponseTime(response.requestOptions.hashCode, DateTime.now());
+    InterceptorManager.instance
+        .onSave(setData(response, response.requestOptions));
     handler.next(response);
   }
 
   @override
   void onError(DioError err, ErrorInterceptorHandler handler) async {
-
     if (err.response != null) {
-      InterceptorManager.instance.onResponseTime(err.response!.requestOptions.hashCode, DateTime.now());
-      InterceptorManager.instance.onSave(setData(err.response!, err.response!.requestOptions)..error = true);
-    }else{
+      InterceptorManager.instance.onResponseTime(
+          err.response!.requestOptions.hashCode, DateTime.now());
+      InterceptorManager.instance.onSave(
+          setData(err.response!, err.response!.requestOptions)..error = true);
+    } else {
       InterceptorManager.instance.onSave(HttpTransaction()
         ..id = err.requestOptions.hashCode
-        ..requestTime = InterceptorManager.instance.requestTime[err.requestOptions.hashCode]
-        ..responseTime = InterceptorManager.instance.responseTime[err.requestOptions.hashCode]
+        ..requestTime =
+            InterceptorManager.instance.requestTime[err.requestOptions.hashCode]
+        ..responseTime = InterceptorManager
+            .instance.responseTime[err.requestOptions.hashCode]
         ..uri = err.requestOptions.uri
         ..error = true);
     }
     handler.next(err);
   }
 
-  HttpTransaction setData(Response response,RequestOptions options){
+  HttpTransaction setData(Response response, RequestOptions options) {
     return HttpTransaction()
       ..id = options.hashCode
       ..uri = options.uri
@@ -47,13 +48,10 @@ class UiNetInterceptor extends Interceptor {
       ..method = options.method
       ..responseType = options.responseType.toString()
       ..followRedirects = options.followRedirects
-      ..connectTimeout = options.connectTimeout
-      ..sendTimeout = options.sendTimeout
-      ..receiveTimeout = options.receiveTimeout
       ..receiveDataWhenStatusError = options.receiveDataWhenStatusError
       // ..extra = convert.jsonEncode(options.extra)
       ..extra = options.extra
-      ..requestBody =  stringifyMessage(options.data)
+      ..requestBody = stringifyMessage(options.data)
       // ..requestHeaders = convert.jsonEncode(options.headers)
       ..requestHeaders = options.headers
       ..redirect = response.isRedirect
@@ -62,8 +60,12 @@ class UiNetInterceptor extends Interceptor {
       ..responseHeaders = response.headers.map
       ..responseBody = response.toString()
       ..requestTime = InterceptorManager.instance.requestTime[options.hashCode]
-      ..responseTime = InterceptorManager.instance.responseTime[options.hashCode]
-      ..duration = InterceptorManager.instance.responseTime[options.hashCode]!.millisecondsSinceEpoch-InterceptorManager.instance.requestTime[options.hashCode]!.millisecondsSinceEpoch;
+      ..responseTime =
+          InterceptorManager.instance.responseTime[options.hashCode]
+      ..duration = InterceptorManager
+              .instance.responseTime[options.hashCode]!.millisecondsSinceEpoch -
+          InterceptorManager
+              .instance.requestTime[options.hashCode]!.millisecondsSinceEpoch;
   }
 
   String stringifyMessage(dynamic message) {
