@@ -5,5 +5,5 @@ import 'package:flutter/services.dart';
 copyClipboard(BuildContext context, String? value) {
   var snackBar = SnackBar(content: Text('copy $value to clipboard'));
   ScaffoldMessenger.of(context).showSnackBar(snackBar);
-  Clipboard.setData(ClipboardData(text: value));
+  Clipboard.setData(ClipboardData(text: value ?? ''));
 }
